backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-chroma
langchain-text-splitters
openai
python-dotenv
docx2txt
PyPDF2
